2011-08-02 Roman - SYS, DAQ: User API functions is appended by Obj.nodePath() and SYS.DAQ["Modul"]["Controller"]["Parameter"]["Attribute"].arch(). - SYS: TiB, PiB, EiB, ZiB and YiB is added to TSYS::cpct2str() function. - Transport.Serial, Transport.Sockets and Transport.SSL: Traffic absolute counter type is changed from float to uint64_t.
git-svn-id: svn://oscada.org/trunk/OpenSCADA@1492 7bc60720-51fb-0310-bb35-cf765b5d350c

+++ src/moduls/transport/SSL/modssl.h
@@ -127,7 +127,7 @@ class TSocketIn: public TTransportIn
 
 	//> Status atributes
 	string		stErr;			// Last error messages
-	float		trIn, trOut;		// Traffic in and out counter
+	uint64_t	trIn, trOut;		// Traffic in and out counter
 	int		connNumb, clsConnByLim;	// Close connections by limit
 };
 
@@ -178,7 +178,7 @@ class TSocketOut: public TTransportOut
 	SSL		*ssl;
 
 	//> Status atributes
-	float		trIn, trOut;		// Traffic in and out counter
+	uint64_t	trIn, trOut;		// Traffic in and out counter
 	Res		wres;
 };
